+++ src/components/base/LanguageSwitcher.vue
@@ -33,6 +33,7 @@ import fr from '../../assets/images/icons/flags/fr.svg'
 import it from '../../assets/images/icons/flags/it.svg'
 import ru from '../../assets/images/icons/flags/ru.svg'
 import jp from '../../assets/images/icons/flags/jp.svg'
+import ko from '../../assets/images/icons/flags/ko.svg'
 import { useI18nStore } from '../../store/i18n'
 import { useTranslation } from '../../composables/i18n.ts'
 import { ValidLocale } from '../../consts.ts'
@@ -59,6 +60,7 @@ const languages: Language[] = [
   { code: 'fr', title: t('base.language_switcher.languages.fr.title'), icon: fr },
   { code: 'it', title: t('base.language_switcher.languages.it.title'), icon: it },
   { code: 'ru', title: t('base.language_switcher.languages.ru.title'), icon: ru },
-  { code: 'jp', title: t('base.language_switcher.languages.jp.title'), icon: jp }
+  { code: 'jp', title: t('base.language_switcher.languages.jp.title'), icon: jp },
+  { code: 'ko', title: t('base.language_switcher.languages.ko.title'), icon: ko }
 ]
 </script>