+++ src/consts.ts
@@ -15,9 +15,11 @@ export const SUPPORTED_COUNTRY_LOCALES: Record<string, ValidLocale> = {
   fr: 'fr',
   it: 'it',
   ru: 'ru',
-  jp: 'jp'
+  jp: 'jp',
+  ko: 'ko',
+  kr: 'ko'
 }
-export type ValidLocale = 'en' | 'cn' | 'fr' | 'it' | 'ru' | 'jp' | 'tw'
+export type ValidLocale = 'en' | 'cn' | 'fr' | 'it' | 'ru' | 'jp' | 'tw' | 'ko'
 export const DEFAULT_LOCALE: ValidLocale = 'en'
 
 export const DEFAULT_DOCUMENT_FIELD_MAX_LENGTH = 200