+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.13.0
 
 * show search aggregations, fixes [#247](https://github.com/cars10/elasticvue/issues/247)
+* i18n: adds traditional chinese, thanks @PeterDaveHello
 
 ## 1.12.0